Add `tzdata` package to `ubuntu` images (#16)
* Add `tzdata` package to `ubuntu` images

This PR adds the `tzdata` package to our `ubuntu` images. `tzdata` is installed by default in the `centos` and `rockylinux` images. The `tzdata` package is necessary for some `arrow` operations, therefore it needs to be installed in our `ubuntu` images as well.

This PR also updates our `condaforge/mambaforge` image to `22.9.0-1`.

* set DEBIAN_FRONTEND=noninteractive for `tzdata` pkg install